+++ demos/AOTAI_Hike/frontend/src/dom.js
@@ -1,0 +1,7 @@
+export const $ = (sel) => document.querySelector(sel);
+
+export const chatEl = $("#chat");
+export const rolesEl = $("#roles");
+export const statusEl = $("#status");
+export const partyEl = $("#party-status");
+export const branchEl = $("#branch-choices");
